fix: clean code & strong type
