+++ packages/app/components/trending/trending.md.tsx
@@ -1,26 +1,21 @@
-import { Suspense, useCallback, useState } from "react";
-import { useWindowDimensions } from "react-native";
+import { Suspense, useCallback, useMemo } from "react";
 
 import { useIsDarkMode } from "@showtime-xyz/universal.hooks";
-import { useRouter } from "@showtime-xyz/universal.router";
 import { Spinner } from "@showtime-xyz/universal.spinner";
 import { Text } from "@showtime-xyz/universal.text";
 import { View } from "@showtime-xyz/universal.view";
 
-import { Card } from "app/components/card";
-import { CreatorPreview } from "app/components/creator-preview";
 import { ErrorBoundary } from "app/components/error-boundary";
 import { TRENDING_ROUTE } from "app/components/trending";
-import { useTrendingCreators, useTrendingNFTS } from "app/hooks/api-hooks";
-import { useContentWidth } from "app/hooks/use-content-width";
-import { InfiniteScrollList } from "app/lib/infinite-scroll-list";
 import { createParam } from "app/navigation/use-param";
 
 import { IndependentTabBar } from "design-system/tab-view/independent-tab-bar";
-import { breakpoints, CARD_DARK_SHADOW } from "design-system/theme";
+
+import { TrendingCreatorsList } from "./trending-creators-list.md";
+import { TrendingNFTSList } from "./trending-nfts-list.md";
 
 type Query = {
-  tab: string;
+  tab: "creator" | "nft";
   days: number;
 };
 
@@ -28,20 +23,13 @@ const { useParam } = createParam<Query>();
 
 export const Trending = () => {
   const [tab, setTab] = useParam("tab");
-  // const isDark = useIsDarkMode();
+  const isDark = useIsDarkMode();
   const selected = tab === "creator" ? 1 : 0;
 
-  // const handleTabChange = useCallback(
-  //   (index: number) => {
-  //     Haptics.impactAsync();
-  //     if (index === 0) {
-  //       setTab("nft");
-  //     } else {
-  //       setTab("creator");
-  //     }
-  //   },
-  //   [setTab]
-  // );
+  const handleTabChange = (index: number) => {
+    // Haptics.impactAsync();
+    setTab(index !== 0 ? "creator" : "nft");
+  };
   const [days, setDays] = useParam("days", {
     initial: 1,
     parse: (value) => Number(value ?? 1),
@@ -74,7 +62,7 @@ export const Trending = () => {
             tw="w-[400px] rounded-lg bg-white p-4 shadow-lg dark:bg-black"
             style={{
               // @ts-ignore
-              boxShadow: isDark ? CARD_DARK_SHADOW : undefined,
+              boxShadow: isDark ? CARD_DARK_SHADOW : CARD_LIGHT_SHADOW,
             }}
           >
             <SegmentedControl
@@ -84,7 +72,6 @@ export const Trending = () => {
             />
           </View> */}
         </View>
-        {/* <TrendingTabs selectedTab={selected === 0 ? "nft" : "creator"} /> */}
         <IndependentTabBar
           onPress={(i) => {
             handleDaysChange(i);
@@ -111,110 +98,17 @@ export const Trending = () => {
   );
 };
 
-const List = ({
-  days,
-  selectedTab,
-}: {
+export type TrendingMDListProps = {
   days: number;
-  selectedTab: "creator" | "nft";
-}) => {
-  if (selectedTab === "creator") {
-    return <CreatorsList days={days} />;
-  }
-  return <NFTList days={days} />;
+  selectedTab?: Query["tab"];
 };
 
-const CreatorsList = ({ days }: { days: any }) => {
-  const { data, isLoading } = useTrendingCreators({
-    days,
-  });
+const LIST_MAP = new Map<Query["tab"], React.FC<TrendingMDListProps>>([
+  ["creator", TrendingCreatorsList],
+  ["nft", TrendingNFTSList],
+]);
 
-  const router = useRouter();
-
-  const [containerWidth, setContainerWidth] = useState(0);
-  const isDark = useIsDarkMode();
-
-  return (
-    <View
-      tw="mt-4 flex-1"
-      onLayout={(e) => setContainerWidth(e.nativeEvent.layout.width)}
-    >
-      {isLoading ? (
-        <View tw="mx-auto p-10">
-          <Spinner />
-        </View>
-      ) : null}
-      {data?.length > 0 && containerWidth
-        ? data.map((item: any) => {
-            return (
-              <View
-                key={item.creator_id}
-                tw="mb-8 rounded-lg bg-white dark:bg-black"
-                //@ts-ignore
-                style={{ boxShadow: isDark ? CARD_DARK_SHADOW : null }}
-              >
-                <CreatorPreview
-                  creator={item}
-                  onMediaPress={(initialScrollIndex: number) => {
-                    router.push(
-                      `/list?initialScrollIndex=${initialScrollIndex}&type=trendingCreator&days=${days}&creatorId=${item.profile_id}`
-                    );
-                  }}
-                  mediaSize={containerWidth / 3 - 2}
-                />
-              </View>
-            );
-          })
-        : null}
-    </View>
-  );
-};
-
-const NFTList = ({ days }: { days: any }) => {
-  const { data, isLoading } = useTrendingNFTS({
-    days,
-  });
-
-  const { width } = useWindowDimensions();
-  const contentWidth = useContentWidth();
-
-  const numColumns = width >= breakpoints["lg"] ? 3 : 2;
-  const renderItem = useCallback(
-    ({ item, index }: any) => {
-      return (
-        <View tw="my-4 flex-1 px-2">
-          <Card
-            hrefProps={{
-              pathname: `/nft/${item.chain_name}/${item.contract_address}/${item.token_id}`,
-            }}
-            key={`nft-list-card-${index}`}
-            nft={item}
-            tw={`w-full h-[${
-              contentWidth / numColumns + 156
-            }px] bg-white dark:bg-black`}
-          />
-        </View>
-      );
-    },
-    [contentWidth, numColumns]
-  );
-  const keyExtractor = useCallback((item: any) => {
-    return item.nft_id;
-  }, []);
-
-  return (
-    <InfiniteScrollList
-      data={data}
-      renderItem={renderItem}
-      numColumns={numColumns}
-      keyExtractor={keyExtractor}
-      ListEmptyComponent={
-        isLoading ? (
-          <View tw="mx-auto p-10">
-            <Spinner />
-          </View>
-        ) : null
-      }
-    />
-  );
+const List = ({ selectedTab = "nft", ...rest }: TrendingMDListProps) => {
+  const List = useMemo(() => LIST_MAP.get(selectedTab), [selectedTab]);
+  return List ? <List {...rest} /> : null;
 };
